MIL. PSR.
Switched to RawF32Ops's "Peak to Sidelobe Ratio"

The custom PSR implementation seem to be working fine.
(1) It has not been checked whether it is equivalent to the approach described
in the paper (probably, won't be)
(2) The experiment shows that when a tracking window loses the target, PSR
shows rather heavy lag
(3) Stable tracking corresponds to PSR in the range [0.45; ).